Added a new test for database consistency just after the installation (number of lines inserted) and moved the tests on the "just installed" database to a standalone phpunit call to ensure they are always called before anything else
